llex_backend/requirements.txt
# ============================================
# LLeX.Ai Backend Dependencies
# ============================================
# Python 3.11+ 권장

# ────────────────────────────────────────────
# 🌐 FastAPI & Web Framework
# ────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# ────────────────────────────────────────────
# 🗄️ Database & ORM
# ────────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0              # PostgreSQL async driver
psycopg2-binary==2.9.10      # PostgreSQL sync driver

# ────────────────────────────────────────────
# 🤖 AI & LLM
# ────────────────────────────────────────────
openai>=1.99.9,<2.0.0
langchain==0.3.27
langchain-core>=0.3.75,<1.0.0
langchain-openai==0.3.32
langgraph==0.2.60

# ────────────────────────────────────────────
# 🧠 Vector Database
# ────────────────────────────────────────────
qdrant-client==1.11.1

# ────────────────────────────────────────────
# 🔧 Configuration & Environment
# ────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.5.2

# ────────────────────────────────────────────
# 🌍 HTTP Clients
# ────────────────────────────────────────────
aiohttp==3.10.11
requests==2.32.3
httpx==0.27.2

# ────────────────────────────────────────────
# 📊 MLOps & Monitoring
# ────────────────────────────────────────────
mlflow==2.18.0
prometheus-client==0.21.0

# ────────────────────────────────────────────
# 🛠️ Utilities
# ────────────────────────────────────────────
rich==13.7.0                 # Beautiful terminal output

# ────────────────────────────────────────────
# 📝 Note
# ────────────────────────────────────────────
# 설치 방법:
#   pip install -r requirements.txt
#
# Docker에서 자동 설치됨 (llex_backend/Dockerfile)
